refactor: rename relay → gateway (remote-gateway-bridge)
Owner-requested naming change, done in the same PR as the feature port:
- src/remote-relay-bridge.py       → src/remote-gateway-bridge.py
- src/remote-relay-bridge.test.py  → src/remote-gateway-bridge.test.py
- docs/remote-relay-protocol.md    → docs/remote-gateway-protocol.md
- terminology in docstrings/comments/log prefix; User-Agent + heartbeat
  client label → sutando-gateway-client
- startup.sh launcher reference updated

Unchanged on purpose: REMOTE_TASK_* env names (never said relay; renaming
them would break every deployed .env) and the /v1 wire protocol.

Compat: a one-release stub remains at src/remote-relay-bridge.py — runs the
renamed client in-process (runpy) so existing launchers and pgrep liveness
checks keep working, with a stderr deprecation line. Suite green under the
new name; stub chain smoke-tested.
