+++ src/remote-gateway-bridge.py
@@ -1,0 +1,597 @@
+#!/usr/bin/env python3
+"""
+remote-gateway-bridge.py — generic client that bridges a REMOTE task gateway to the
+local Sutando file queue, so the local core processes remote tasks unchanged.
+
+This is the OPEN, provider-agnostic half of the "agent as a service" design: a
+gateway service holds the platform connection and
+exposes a tiny HTTP protocol; this client pulls *your* tasks down into the local
+`tasks/` queue and pushes results back up. No provider-specific logic lives here
+— that's the gateway's job.
+
+Full spec: docs/remote-gateway-protocol.md
+
+  Protocol (versioned, Bearer-auth):
+  GET  {REMOTE_TASK_URL}/v1/tasks?wait=<sec>
+       → 200 {"tasks": [ {<task fields...>}, ... ]}   (long-poll; [] on timeout)
+  POST {REMOTE_TASK_URL}/v1/tasks/<task-id>/ack
+       → body {"id": "<task-id>"}  → 200 on accepted
+  POST {REMOTE_TASK_URL}/v1/results
+       → body {"id": "<task-id>", "body": "<result text>"}  → 200 on accepted
+  POST {REMOTE_TASK_URL}/v1/heartbeat
+       → body {"client": "...", "inflight": N, ...}  → 200 on accepted
+
+Each task object uses the same schema Sutando's other bridges write, so this
+client just serializes it to `tasks/task-<id>.txt` and the core handles it like
+any Discord/Telegram/Slack task. When `results/task-<id>.txt` appears, its body
+is POSTed back and the result file is archived. Ack/heartbeat are best-effort:
+if an older gateway returns 404/405, the client keeps working against the
+original pull/result protocol.
+
+Config (env / .env):
+  REMOTE_TASK_TOKEN      the onboarding string — the ONLY required setting
+                        (combined "https://<gateway>|<secret>" or a bare secret)
+  REMOTE_TASK_URL        gateway base URL (only needed with a bare secret)
+  REMOTE_TASK_URL/_TOKEN  legacy aliases
+  REMOTE_TASK_PROVIDER  label used for the task `source:` field (default "remote")
+  REMOTE_TASK_POLL_WAIT long-poll seconds (default 25)
+
+Stdlib only (urllib) — no new dependencies.
+"""
+
+from __future__ import annotations
+
+import json
+import os
+import re
+import sys
+import time
+import urllib.error
+import urllib.parse
+import urllib.request
+from pathlib import Path
+
+# resolve_workspace lives alongside this file in src/.
+# Coupled-skill import: this skill ships in the main repo, so use the
+# canonical src/ helper rather than a vendored copy (avoids silent drift).
+sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "src"))
+from workspace_default import resolve_workspace  # noqa: E402
+
+WS = resolve_workspace()
+TASKS_DIR = WS / "tasks"
+RESULTS_DIR = WS / "results"
+ARCHIVE_RESULTS_DIR = RESULTS_DIR / "archive"
+# Persist the in-flight set (tasks pulled from the gateway, awaiting result-POST)
+# so a client restart between pull and POST doesn't strand the result. Scoped to
+# gateway-pulled tasks only — we must NOT blindly POST every results/ file, or we'd
+# cross-send other channels' (Discord/Telegram) results to the gateway.
+INFLIGHT_FILE = WS / "state" / "remote-task-inflight.json"
+
+# Back-compat: instances onboarded before the AG2_REMOTE_* → REMOTE_TASK_*
+# rename still export the legacy names in their .env. Honor them as DEPRECATED
+# aliases for one release (remove next), with a one-line migration nudge, so the
+# bridge keeps connecting under any launcher. New onboards use REMOTE_TASK_*.
+_warned_legacy = set()
+def _env_compat(new, old):
+    v = os.environ.get(new)
+    if v:
+        return v
+    v = os.environ.get(old)
+    if v and old not in _warned_legacy:
+        _warned_legacy.add(old)
+        print(f"[remote-gateway-bridge] {old} is deprecated — rename to {new} in your .env",
+              file=sys.stderr, flush=True)
+    return v
+
+# One-token onboarding: REMOTE_TASK_TOKEN alone is enough. The onboarding
+# string may be the combined "https://<gateway>|<secret>" form (the URL travels
+# inside the token — nothing service-specific lives in this repo); a bare
+# secret needs REMOTE_TASK_URL alongside it.
+_RAW = _env_compat("REMOTE_TASK_TOKEN", "AG2_REMOTE_TOKEN") or ""
+if "|" in _RAW:
+    _URL_FROM_TOKEN, TOKEN = _RAW.split("|", 1)
+else:
+    _URL_FROM_TOKEN, TOKEN = "", _RAW
+URL = (_env_compat("REMOTE_TASK_URL", "AG2_REMOTE_URL")
+       or _URL_FROM_TOKEN).rstrip("/")
+PROVIDER = os.environ.get("REMOTE_TASK_PROVIDER") or "remote"
+POLL_WAIT = int(os.environ.get("REMOTE_TASK_POLL_WAIT") or "25")
+HEARTBEAT_INTERVAL = 60
+_ack_disabled = False
+_heartbeat_disabled = False
+_last_heartbeat_at = 0.0
+
+_TASK_FIELDS = ("id", "timestamp", "task", "source", "channel_id",
+                # Context enrichment (AG2 broker writer side): human room/sender
+                # names + reply reference. Serialized only when the gateway sends
+                # them (absent for other sources); each newline-stripped by
+                # _one_line so a room/display name can't forge an extra line.
+                "room_name", "sender_name", "reply_to_event", "reply_to_me",
+                "source_message_id", "user_id", "priority")
+
+# Trust tier is a LOCAL decision (review 2026-06-13): the gateway is outside
+# this machine's trust boundary, so its access_tier claim is ignored. The
+# tier written to every task file comes from REMOTE_TASK_TIER in .env —
+# default "team" (sandboxed processing). Operators who own their gateway can
+# explicitly set REMOTE_TASK_TIER=owner.
+LOCAL_TIER = (_env_compat("REMOTE_TASK_TIER", "AG2_REMOTE_TIER") or "team").strip().lower()
+if LOCAL_TIER not in ("owner", "team", "other"):
+    LOCAL_TIER = "team"
+
+# ── inbound media fetch (owner screenshots, file uploads) ────────────────────
+# A gateway can hand the task body a media MARKER instead of raw bytes:
+#   [<tag>: <url> mime=<m> name=<f> size=<n> kind=<msgtype>] <caption>
+# `<url>` is typically unreachable for the core as-is (a homeserver media URL
+# behind authenticated-media, or a gateway media-proxy URL). We resolve it here —
+# where the gateway bearer already lives — download the bytes to a local file,
+# and rewrite the marker to `[File attached: <path>]` (the inbound convention
+# the Discord/Telegram bridges use) so the core just reads a local path with
+# zero remote creds.
+#
+# Auth is picked by the URL:
+#   • URL under REMOTE_TASK_URL → fetched with the gateway bearer we already hold
+#   • a Matrix `/_matrix/media/...` URL → upgraded to the authenticated
+#     MSC3916 client route and fetched with REMOTE_MEDIA_HS_TOKEN (a homeserver
+#     access token), if configured
+#   • any other https URL → fetched with NO credentials
+# Authenticated fetches do NOT follow redirects (a gateway-controlled URL must
+# not be able to bounce our bearer to a third-party host). Drop-in safe: no
+# token / fetch error / oversize → the marker is left untouched.
+#
+# The marker tag is configurable (REMOTE_MEDIA_MARKER, slug chars only) so a
+# provider-specific gateway can keep its existing marker name without this repo
+# carrying provider strings.
+MEDIA_MARKER_TAG = re.sub(r"[^A-Za-z0-9_-]", "",
+                          os.environ.get("REMOTE_MEDIA_MARKER") or "remote-media")
+MEDIA_MARKER_RE = re.compile(r"\[" + re.escape(MEDIA_MARKER_TAG) + r":([^\]]*)\]")
+HS_MEDIA_TOKEN = os.environ.get("REMOTE_MEDIA_HS_TOKEN") or ""
+MEDIA_DIR = Path(os.environ.get("REMOTE_MEDIA_DIR") or str(WS / "data" / "remote-media"))
+MAX_MEDIA_BYTES = int(os.environ.get("REMOTE_MEDIA_MAX_BYTES") or str(25 * 1024 * 1024))
+_EXT_BY_MIME = {
+    "image/png": ".png", "image/jpeg": ".jpg", "image/jpg": ".jpg",
+    "image/gif": ".gif", "image/webp": ".webp", "image/svg+xml": ".svg",
+    "application/pdf": ".pdf",
+}
+
+# Bridges-as-siblings: discord/telegram/slack bridges write
+# `state/last-owner-activity.json` whenever the owner messages them, so the
+# proactive-loop's "active engagement" gate knows a conversation is live. The
+# gateway transport should feed the same gate — but only when THIS node treats
+# gateway traffic as owner traffic (LOCAL_TIER, never the gateway's claim).
+OWNER_ACTIVITY_FILE = WS / "state" / "last-owner-activity.json"
+
+
+# Blocker (review 2026-06-13): the gateway is untrusted, so a task `id` flows
+# into filesystem paths (task write + result read-back/POST). Reject anything
+# that isn't a plain slug — kills path traversal in both directions.
+_TID_RE = re.compile(r"[A-Za-z0-9._-]{1,64}")
+
+
+def _valid_tid(tid: str) -> bool:
+    return bool(_TID_RE.fullmatch(tid)) and tid not in (".", "..")
+
+
+def _one_line(value) -> str:
+    """Header-safe single-line value: CR/LF stripped so a gateway-controlled
+    field can't inject extra `key: value` lines (e.g. forge a second
+    access_tier). Applied to every field — task content is single-line in
+    practice and a stray newline only ever indicates an injection attempt."""
+    return str(value).replace("\r", " ").replace("\n", " ")
+
+
+def _log(msg: str) -> None:
+    print(f"[remote-gateway-bridge] {msg}", flush=True)
+
+
+def _req(method: str, path: str, payload: dict | None = None, timeout: int = 35):
+    """One authenticated HTTP request. Returns parsed JSON (or {} for empty)."""
+    data = json.dumps(payload).encode() if payload is not None else None
+    req = urllib.request.Request(f"{URL}{path}", data=data, method=method)
+    req.add_header("Authorization", f"Bearer {TOKEN}")
+    req.add_header("Accept", "application/json")
+    # CloudFlare bot-fight (error 1010) rejects python-urllib's default
+    # User-Agent with a 403; send an explicit client UA so the gateway's edge
+    # lets the long-poll through. (Same fix the other gateway callers carry.)
+    req.add_header("User-Agent", "sutando-gateway-client/1.0")
+    if data is not None:
+        req.add_header("Content-Type", "application/json")
+    with urllib.request.urlopen(req, timeout=timeout) as resp:
+        raw = resp.read().decode().strip()
+        return json.loads(raw) if raw else {}
+
+
+def _post_task_ack(tid: str) -> bool:
+    """Tell the gateway a task made it safely into the local queue."""
+    global _ack_disabled
+    if _ack_disabled or not _valid_tid(tid):
+        return False
+    try:
+        safe_tid = urllib.parse.quote(tid, safe="")
+        _req("POST", f"/v1/tasks/{safe_tid}/ack", {"id": tid}, timeout=10)
+        return True
+    except urllib.error.HTTPError as e:
+        if e.code in (404, 405):
+            _ack_disabled = True
+            _log("gateway does not support task ack — continuing without")
+            return False
+        if e.code in (401, 403):
+            raise
+        _log(f"task ack failed for {tid}: HTTP {e.code} — gateway may redeliver")
+        return False
+    except (urllib.error.URLError, TimeoutError) as e:
+        _log(f"task ack network error for {tid}: {e} — gateway may redeliver")
+        return False
+
+
+_CORE_STEP_MAX = 500
+
+
+def _core_str(v) -> str | None:
+    """A core-status field → bounded non-empty str, or None. core-status.json is
+    written by another process and may be malformed; a non-string field must not
+    be forwarded (the broker calls .lower() on it)."""
+    if not isinstance(v, str):
+        return None
+    v = v.strip()
+    return v[:_CORE_STEP_MAX] if v else None
+
+
+def _read_core_status() -> tuple[str | None, str | None]:
+    """Read this node's core-status.json → (status, step) for the presence layer.
+    core-status is written by the proactive loop / task handlers (status =
+    running|idle, step = human 'what it's doing'). The broker derives the agent's
+    presence badge from it.
+
+    MUST NOT raise: this runs in the main loop BEFORE the /v1/tasks poll, so an
+    exception here would back the loop off and stall task delivery — a malformed
+    presence side-channel must never become a delivery blocker. So we guard the
+    JSON shape (a valid-JSON non-object would AttributeError on .get) and coerce
+    every field to a bounded str-or-None; any surprise → (None, None) and the
+    heartbeat still fires as a plain liveness ping."""
+    try:
+        with open(WS / "state" / "core-status.json") as f:
+            cs = json.load(f)
+        if not isinstance(cs, dict):
+            return (None, None)
+        status = _core_str(cs.get("status"))
+        step = _core_str(cs.get("step"))
+        # An idle status carries no meaningful step — send status only so the
+        # sweep reads 'available' rather than stale 'what it was last doing'.
+        return (status, None if status == "idle" else step)
+    except Exception:  # noqa: BLE001 — best-effort; never stall the main loop
+        return (None, None)
+
+
+def _post_heartbeat(inflight: set[str], force: bool = False) -> bool:
+    """Best-effort liveness + core-status ping. Liveness feeds hosted dashboards;
+    the status/step feed the broker's presence sweep (agent working/available/…)."""
+    global _heartbeat_disabled, _last_heartbeat_at
+    if _heartbeat_disabled:
+        return False
+    now = time.time()
+    if not force and now - _last_heartbeat_at < HEARTBEAT_INTERVAL:
+        return False
+    _last_heartbeat_at = now
+    _status, _step = _read_core_status()
+    try:
+        payload = {
+            "client": "sutando-gateway-client",
+            "protocol_version": 1,
+            "provider": PROVIDER,
+            "tier": LOCAL_TIER,
+            "inflight": len(inflight),
+            "capabilities": ["task-ack", "heartbeat", "result-skip-markers",
+                             "core-status"],
+        }
+        # Only include when present so a status-less node never clobbers the
+        # broker's last-known core-status (the broker only records on presence).
+        if _status is not None:
+            payload["status"] = _status
+        if _step is not None:
+            payload["step"] = _step
+        _req("POST", "/v1/heartbeat", payload, timeout=10)
+        return True
+    except urllib.error.HTTPError as e:
+        if e.code in (404, 405):
+            _heartbeat_disabled = True
+            _log("gateway does not support heartbeat — continuing without")
+            return False
+        if e.code in (401, 403):
+            raise
+        _log(f"heartbeat failed: HTTP {e.code} — continuing")
+        return False
+    except (urllib.error.URLError, TimeoutError) as e:
+        _log(f"heartbeat network error: {e} — continuing")
+        return False
+
+
+def _marker_attr(attrs: str, key: str) -> str:
+    """Pull a `key=value` (value = non-space run) out of the marker attr tail."""
+    m = re.search(rf"\b{re.escape(key)}=([^\s\]]+)", attrs)
+    return m.group(1) if m else ""
+
+
+def _to_authed_media_url(url: str) -> str:
+    """Upgrade a legacy unauthenticated Matrix media route to the MSC3916
+    authenticated client route (Matrix v1.11+). Leaves other URLs untouched.
+      /_matrix/media/(r0|v3)/download/<server>/<id>
+        → /_matrix/client/v1/media/download/<server>/<id>"""
+    return re.sub(r"/_matrix/media/(?:r0|v3)/download/",
+                  "/_matrix/client/v1/media/download/", url, count=1)
+
+
+def _download_bytes(url: str, headers: dict, cap: int) -> bytes:
+    """GET raw bytes with an explicit size cap (reads cap+1 then rejects if
+    over, so a missing/lying Content-Length can't OOM us). When an
+    Authorization header is present, redirects are NOT followed — a
+    gateway-controlled URL must not bounce our bearer to another host."""
+    req = urllib.request.Request(url, method="GET")
+    for k, v in headers.items():
+        req.add_header(k, v)
+    if "Authorization" in headers:
+        class _NoRedirect(urllib.request.HTTPRedirectHandler):
+            def redirect_request(self, *a, **kw):  # noqa: D401
+                return None
+        opener = urllib.request.build_opener(_NoRedirect)
+        resp_ctx = opener.open(req, timeout=30)
+    else:
+        resp_ctx = urllib.request.urlopen(req, timeout=30)
+    with resp_ctx as resp:
+        data = resp.read(cap + 1)
+    if len(data) > cap:
+        raise ValueError(f"media exceeds {cap}-byte cap")
+    return data
+
+
+def _maybe_fetch_media(body: str) -> str:
+    """If `body` carries a media marker, download the attachment to a local
+    file and rewrite the marker to `[File attached: <path>]`. Returns the body
+    unchanged on any failure (drop-in safe)."""
+    m = MEDIA_MARKER_RE.search(body or "")
+    if not m:
+        return body
+    inner = m.group(1).strip()
+    if not inner:
+        return body
+    parts = inner.split(None, 1)
+    url = parts[0]
+    attrs = parts[1] if len(parts) > 1 else ""
+    mime = _marker_attr(attrs, "mime")
+    name = _marker_attr(attrs, "name")
+    kind = _marker_attr(attrs, "kind")
+
+    if not url.startswith(("https://", "http://")):
+        return body
+    headers = {"User-Agent": "sutando-gateway-client/1.0"}
+    if URL and url.startswith(URL):
+        headers["Authorization"] = f"Bearer {TOKEN}"            # gateway media-proxy
+    elif "/_matrix/" in url:
+        if not HS_MEDIA_TOKEN:
+            return body                                          # can't auth — leave marker
+        url = _to_authed_media_url(url)
+        headers["Authorization"] = f"Bearer {HS_MEDIA_TOKEN}"
+    # else: a plain public URL — fetched with no credentials.
+
+    try:
+        data = _download_bytes(url, headers, MAX_MEDIA_BYTES)
+    except Exception as e:  # noqa: BLE001 — drop-in safe
+        _log(f"media fetch failed ({e}) — leaving marker as-is")
+        return body
+    try:
+        MEDIA_DIR.mkdir(parents=True, exist_ok=True)
+        ext = _EXT_BY_MIME.get(mime.lower(), "")
+        if not ext and "." in name:
+            ext = "." + re.sub(r"[^A-Za-z0-9]", "", name.rsplit(".", 1)[1])[:8]
+        safe = re.sub(r"[^A-Za-z0-9._-]", "_", name) or "attachment"
+        if ext and safe.endswith(ext):
+            safe = safe[: -len(ext)]
+        path = MEDIA_DIR / f"{int(time.time() * 1000)}-{safe}{ext}"
+        path.write_bytes(data)
+    except Exception as e:  # noqa: BLE001
+        _log(f"media save failed ({e}) — leaving marker as-is")
+        return body
+    _log(f"fetched media → {path} ({len(data)} bytes)")
+    label = "Photo attached" if str(kind) == "m.image" else "File attached"
+    return MEDIA_MARKER_RE.sub(f"[{label}: {path}]", body, count=1)
+
+
+def _write_owner_activity(task: dict) -> None:
+    """Record that the owner was active on this transport right now — but only
+    when THIS node grants gateway traffic owner tier (LOCAL_TIER, never the
+    gateway's own claim; the gateway is outside the trust boundary). Atomic write
+    via tmp+rename; same schema (`ts`, `channel`, `summary`) as
+    discord-bridge.write_owner_activity so the proactive-loop reader is
+    transport-agnostic. Best-effort — never blocks task intake."""
+    if LOCAL_TIER != "owner":
+        return
+    try:
+        OWNER_ACTIVITY_FILE.parent.mkdir(parents=True, exist_ok=True)
+        # Strip a bracket prefix a gateway may add (e.g. "[Provider @user] body").
+        body = (task.get("task") or "").lstrip()
+        if body.startswith("[") and "]" in body:
+            body = body[body.index("]") + 1:].lstrip()
+        payload = {
+            "ts": int(time.time()),
+            "channel": task.get("source") or PROVIDER,
+            "summary": body[:80],
+        }
+        tmp = OWNER_ACTIVITY_FILE.with_suffix(".json.tmp")
+        tmp.write_text(json.dumps(payload))
+        tmp.rename(OWNER_ACTIVITY_FILE)
+    except Exception as e:  # noqa: BLE001
+        _log(f"owner-activity write failed: {e}")
+
+
+def _write_task(task: dict) -> str | None:
+    """Serialize a gateway task into tasks/task-<id>.txt (same schema as bridges).
+    Returns the task id, or None if it has no id / already present."""
+    tid = str(task.get("id") or "").strip()
+    if not tid:
+        _log("dropping task with no id")
+        return None
+    if not _valid_tid(tid):
+        _log(f"dropping task with unsafe id {tid!r}")
+        return None
+    dest = TASKS_DIR / f"{tid}.txt"
+    # Idempotent: don't re-write a task already queued, claimed, or archived.
+    if dest.exists() or any(TASKS_DIR.glob(f"{tid}.claimed-*")):
+        return tid
+    # Relay redelivery of already-handled work: on reconnect the gateway replays
+    # its unacked pool, including tasks this node long since processed (the
+    # 2026-06-30 and 2026-07-01 500-task floods). If the core already archived
+    # the task file, or the result was already delivered and archived, don't
+    # re-queue — drop a [no-send] result instead so the normal result drain
+    # re-acks it upstream and clears it from inflight.
+    _task_archive = TASKS_DIR / "archive"
+    task_archived = (
+        # legacy flat layout: tasks/archive/<taskId>.txt
+        (_task_archive / f"{tid}.txt").exists()
+        # active month-partitioned layout: tasks/archive/YYYY-MM/<taskId>.txt
+        # (see src/task-bridge.ts). Glob one level of month subdirs for this
+        # exact task id — cheap (one stat per month dir, not a full tree walk).
+        or next(_task_archive.glob(f"*/{tid}.txt"), None) is not None
+    )
+    if (task_archived
+            or (ARCHIVE_RESULTS_DIR / f"{tid}.txt").exists()
+            or next(ARCHIVE_RESULTS_DIR.glob(f"{tid}-[0-9]*.txt"), None)):
+        rfile = RESULTS_DIR / f"{tid}.txt"
+        if not rfile.exists():
+            RESULTS_DIR.mkdir(parents=True, exist_ok=True)
+            rfile.write_text("[no-send] gateway redelivery of already-handled task\n")
+        _log(f"dedup: {tid} already handled — not re-queued")
+        return tid
+    TASKS_DIR.mkdir(parents=True, exist_ok=True)
+    lines = []
+    for f in _TASK_FIELDS:
+        if f == "source":
+            lines.append(f"source: {_one_line(task.get('source') or PROVIDER)}")
+        elif f == "task" and task.get("task") not in (None, ""):
+            # Resolve an inbound media marker to a local file the core can read.
+            lines.append(f"task: {_one_line(_maybe_fetch_media(str(task['task'])))}")
+        elif f in task and task[f] not in (None, ""):
+            lines.append(f"{f}: {_one_line(task[f])}")
+    # access_tier is a LOCAL decision and written LAST so it wins even under a
+    # last-occurrence parser; every other field is newline-stripped so none can
+    # forge an earlier one either.
+    lines.append(f"access_tier: {LOCAL_TIER}")
+    tmp = dest.with_suffix(".txt.tmp")
+    tmp.write_text("\n".join(lines) + "\n")
+    tmp.rename(dest)  # atomic publish so the watcher never sees a partial file
+    _log(f"queued {tid}")
+    # Bridges-as-siblings: feed the proactive-loop's active-engagement gate.
+    _write_owner_activity(task)
+    return tid
+
+
+def _archive_result(path: Path, tid: str) -> None:
+    ARCHIVE_RESULTS_DIR.mkdir(parents=True, exist_ok=True)
+    try:
+        path.rename(ARCHIVE_RESULTS_DIR / f"{tid}-{int(time.time())}.txt")
+    except OSError:
+        path.unlink(missing_ok=True)
+
+
+def _load_inflight() -> set[str]:
+    """Restore the in-flight set from disk (fail-open to empty)."""
+    try:
+        data = json.loads(INFLIGHT_FILE.read_text())
+        return {str(t) for t in data} if isinstance(data, list) else set()
+    except FileNotFoundError:
+        return set()
+    except Exception as e:  # noqa: BLE001
+        _log(f"inflight file unreadable ({e}) — starting empty")
+        return set()
+
+
+def _save_inflight(inflight: set[str]) -> None:
+    """Atomically persist the in-flight set. Best-effort (never blocks the loop)."""
+    try:
+        INFLIGHT_FILE.parent.mkdir(parents=True, exist_ok=True)
+        tmp = INFLIGHT_FILE.with_suffix(".json.tmp")
+        tmp.write_text(json.dumps(sorted(inflight)))
+        tmp.rename(INFLIGHT_FILE)
+    except Exception as e:  # noqa: BLE001
+        _log(f"inflight persist failed ({e}) — continuing")
+
+
+def _post_ready_results(inflight: set[str]) -> None:
+    """For each in-flight task, if its result file exists, POST it + archive."""
+    changed = False
+    for tid in list(inflight):
+        if not _valid_tid(tid):  # defense-in-depth: never read an unsafe path
+            inflight.discard(tid); changed = True
+            continue
+        rfile = RESULTS_DIR / f"{tid}.txt"
+        if not rfile.exists():
+            continue
+        body = rfile.read_text().strip()
+        # Result-body protocol markers are a local bridge concern — never ship
+        # them to the gateway. [no-send]/[deduped:] mean "no user-facing reply":
+        # archive without POSTing (match the other bridges' semantics).
+        low = body.lower()
+        if low.startswith("[no-send]") or low.startswith("[deduped:"):
+            _archive_result(rfile, tid)
+            inflight.discard(tid)
+            changed = True
+            _log(f"archived {tid} (marker, not sent)")
+            continue
+        try:
+            _req("POST", "/v1/results", {"id": tid, "body": body})
+        except urllib.error.HTTPError as e:
+            _log(f"result POST failed for {tid}: HTTP {e.code} — will retry")
+            continue
+        except (urllib.error.URLError, TimeoutError) as e:
+            _log(f"result POST network error for {tid}: {e} — will retry")
+            continue
+        _archive_result(rfile, tid)
+        inflight.discard(tid)
+        changed = True
+        _log(f"delivered result for {tid}")
+    if changed:
+        _save_inflight(inflight)
+
+
+def main() -> None:
+    if not URL or not TOKEN:
+        sys.exit("FATAL: set REMOTE_TASK_TOKEN (and REMOTE_TASK_URL if your token is a bare secret).")
+    inflight: set[str] = _load_inflight()
+    _log(f"starting — gateway={URL} provider={PROVIDER} workspace={WS} "
+         f"(restored {len(inflight)} in-flight)")
+    backoff = 1
+    while True:
+        try:
+            _post_heartbeat(inflight)
+            resp = _req("GET", f"/v1/tasks?wait={POLL_WAIT}", timeout=POLL_WAIT + 10)
+            added = False
+            pending_ack = []
+            for task in resp.get("tasks", []):
+                tid = _write_task(task)
+                if tid:
+                    if tid not in inflight:
+                        inflight.add(tid)
+                        added = True
+                    pending_ack.append(tid)
+            if added:
+                _save_inflight(inflight)
+            # Ack only after both the task file and local in-flight state are
+            # durable, so a crash after ack does not strand the eventual result.
+            for tid in pending_ack:
+                _post_task_ack(tid)
+            _post_ready_results(inflight)
+            _post_heartbeat(inflight)
+            backoff = 1  # healthy round-trip → reset backoff
+        except urllib.error.HTTPError as e:
+            if e.code in (401, 403):
+                sys.exit(f"FATAL: gateway auth rejected (HTTP {e.code}) — check REMOTE_TASK_TOKEN.")
+            _log(f"poll HTTP {e.code} — backing off {backoff}s")
+            time.sleep(backoff); backoff = min(backoff * 2, 60)
+        except (urllib.error.URLError, TimeoutError) as e:
+            _log(f"poll network error: {e} — backing off {backoff}s")
+            time.sleep(backoff); backoff = min(backoff * 2, 60)
+        except Exception as e:  # noqa: BLE001 — keep the loop alive
+            _log(f"unexpected: {e} — backing off {backoff}s")
+            time.sleep(backoff); backoff = min(backoff * 2, 60)
+
+
+if __name__ == "__main__":
+    main()
